api/temp_requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic[email]==2.7.1
pydantic-settings==2.2.1
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
alembic==1.13.1
celery[redis]==5.4.0
redis==5.0.4
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
boto3==1.34.100
httpx==0.27.0
pytest==8.2.0
pytest-asyncio==0.23.6
stripe==9.8.0
resend==0.7.2
python-dotenv==1.0.1

# Phase 3 — AI / cheminformatics
ruff==0.4.4                # linter (also used by CI)
biopython==1.83            # CIF → PDB conversion for AlphaFold Server outputs
